refactor(Card, ImageSlider): make favorites icon from Card and placed it into ImageSlider
