+++ src/components/Card.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { FiHeart } from 'react-icons/fi';
 // import EffectSlider from './EffectSlider';
 // import {
 //   useAppDispatch,
@@ -13,32 +12,25 @@ import { FiHeart } from 'react-icons/fi';
 import styles from '../../styles/ImageGallery.module.css';
 import ImageSlider from './ImageSlider';
 
-const Card = ({image}) => {
+const Card = ({ image }) => {
   // const dispatch = useAppDispatch();
-  const [isFavorite, setFavorite] = useState(false);
+  // const [isFavorite, setFavorite] = useState(false);
   // const { favorites } = useAppSelector(selectFavorites);
 
-  const handleFavorite = () => {
-    setFavorite(!isFavorite);
-    // dispatch(addToFavorites('image'));
-    // dispatch(removeFavorite('image'));
-  };
+  // const handleFavorite = () => {
+  //   setFavorite(!isFavorite);
+  //   dispatch(addToFavorites('image'));
+  //   dispatch(removeFavorite('image'));
+  // };
 
   // console.log(favorites);
 
   return (
-    <div className={styles.card}>
-      <div className={styles.imageContainer}>
-        <div className={styles.favoriteIcon}>
-          <FiHeart
-            onClick={handleFavorite}
-            fill={isFavorite ? 'grey' : 'white'}
-            size="11"
-          />
-        </div>
-        <div className={styles.imageSlider}>
-          <ImageSlider image={image} />
-        </div>
+    <div className={`${styles.card}`}>
+      <div>
+        <ImageSlider
+          image={image}
+        />
       </div>
     </div>
   );
