+++ src/components/ImageSlider.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import { FiHeart } from 'react-icons/fi';
 import Image from 'next/image';
 import {
   Slider,
@@ -9,13 +10,26 @@ import {
 } from '@chakra-ui/react';
 import styles from '../../styles/ImageGallery.module.css';
 
-const ImageSlider = ({image}) => {
+const ImageSlider = ({ image }) => {
+
+  const [isFavorite, setFavorite] = useState(false);
+
+  const handleFavorite = () => {
+    setFavorite(!isFavorite);
+  };
   const [sliderValue, setSliderValue] = useState(5);
   const [sliderRange, setSliderRange] = useState('0%');
   const [showTooltip, setShowTooltip] = useState(false);
-  // console.log(image.src.original);
+ 
   return (
-    <>
+    <div className={styles.imageContainer}>
+      <div className={styles.favoriteIcon}>
+        <FiHeart
+          onClick={handleFavorite}
+          fill={isFavorite ? 'grey' : 'white'}
+          size="11"
+        />
+      </div>
       <Image
         style={{ filter: `grayscale(${sliderRange})` }}
         className={styles.image}
@@ -50,7 +64,7 @@ const ImageSlider = ({image}) => {
           <SliderThumb />
         </Tooltip>
       </Slider>
-    </>
+    </div>
   );
 };
