+++ styles/ImageGallery.module.css
@@ -97,4 +97,3 @@
     overflow:scroll;
     height: 100px;
 } */
-
